Added "groups" (RUN NEW SQL FILE!)
each user now has a "groups" entry in the db, giving a list of
subscribed group ids. these show up on the side menu thing. on line 128
of home.php, there is a function that needs to get the name of a group
in the database from $id, and return it.